backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
aiomysql
python-dotenv
alembic
pydantic-settings
PyMySQL
email-validator
cryptography
python-jose[cryptography]
python-multipart